Move SignedPacket into a new crate iroh-dns
This allows us to not depend on iroh-relay in iroh-dns-server
