+++ iroh-dns-server/Cargo.toml
@@ -30,6 +30,7 @@ humantime = "2.2.0"
 humantime-serde = "1.1.1"
 iroh-base = { version = "0.97", path = "../iroh-base", features = ["key"] }
 iroh-metrics = { version = "0.38", features = ["service"] }
+iroh-dns = { version = "0.97", path = "../iroh-dns" }
 iroh-relay = { version = "0.97", path = "../iroh-relay", default-features = false }
 lru = "0.16.3"
 mainline = "6"
