+++ iroh-dns/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "iroh-dns"
+version = "0.97.0"
+edition = "2024"
+description = "DNS-based endpoint discovery for iroh"
+license = "MIT OR Apache-2.0"
+authors = ["n0 team"]
+repository = "https://github.com/n0-computer/iroh"
+keywords = ["networking", "pkarr", "dns", "iroh"]
+rust-version = "1.89"
+
+[lints]
+workspace = true
+
+[dependencies]
+data-encoding = "2.6.0"
+iroh-base = { version = "0.97.0", path = "../iroh-base", default-features = false, features = ["key"] }
+n0-error = "0.1"
+n0-future = "0.3"
+simple-dns = "0.9"
